Dərslik/testgen/backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.49
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
PyJWT[crypto]==2.10.1
bcrypt==4.3.0
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.28.1
google-genai==1.68.0
qdrant-client==1.17.1
python-dotenv==1.0.1
reportlab==4.4.0
matplotlib==3.10.8
python-docx==1.1.2
pymupdf==1.25.0
pytest==8.3.4
pytest-asyncio==0.24.0